Refactored uid utils
