+++ packages/strapi-plugin-content-manager/test/content-manager/uid.test.e2e.js
@@ -39,7 +39,7 @@ describe('Content Manager single types', () => {
   describe('Generate UID', () => {
     test('Throws if input is not provided', async () => {
       const res = await rq({
-        url: `/content-manager/explorer/uid/generate`,
+        url: `/content-manager/uid/generate`,
         method: 'POST',
         body: {},
       });
@@ -59,7 +59,7 @@ describe('Content Manager single types', () => {
 
     test('Throws when contentType is not found', async () => {
       const res = await rq({
-        url: `/content-manager/explorer/uid/generate`,
+        url: `/content-manager/uid/generate`,
         method: 'POST',
         body: {
           contentTypeUID: 'non-existent',
@@ -79,7 +79,7 @@ describe('Content Manager single types', () => {
 
     test('Throws when field is not a uid field', async () => {
       const res = await rq({
-        url: `/content-manager/explorer/uid/generate`,
+        url: `/content-manager/uid/generate`,
         method: 'POST',
         body: {
           contentTypeUID: uid,
@@ -99,9 +99,9 @@ describe('Content Manager single types', () => {
       });
     });
 
-    test('Generates a unique field when not targetField', async () => {
+    test('Generates a unique field when targetField is empty', async () => {
       const res = await rq({
-        url: `/content-manager/explorer/uid/generate`,
+        url: `/content-manager/uid/generate`,
         method: 'POST',
         body: {
           contentTypeUID: uid,
@@ -122,7 +122,7 @@ describe('Content Manager single types', () => {
       });
 
       const secondRes = await rq({
-        url: `/content-manager/explorer/uid/generate`,
+        url: `/content-manager/uid/generate`,
         method: 'POST',
         body: {
           contentTypeUID: uid,
@@ -137,7 +137,7 @@ describe('Content Manager single types', () => {
 
     test('Generates a unique field based on targetField', async () => {
       const res = await rq({
-        url: `/content-manager/explorer/uid/generate`,
+        url: `/content-manager/uid/generate`,
         method: 'POST',
         body: {
           contentTypeUID: uid,
@@ -160,7 +160,7 @@ describe('Content Manager single types', () => {
       });
 
       const secondRes = await rq({
-        url: `/content-manager/explorer/uid/generate`,
+        url: `/content-manager/uid/generate`,
         method: 'POST',
         body: {
           contentTypeUID: uid,
@@ -177,7 +177,7 @@ describe('Content Manager single types', () => {
 
     test('Avoids collisions with already generated uids', async () => {
       const res = await rq({
-        url: `/content-manager/explorer/uid/generate`,
+        url: `/content-manager/uid/generate`,
         method: 'POST',
         body: {
           contentTypeUID: uid,
@@ -200,7 +200,7 @@ describe('Content Manager single types', () => {
       });
 
       const secondRes = await rq({
-        url: `/content-manager/explorer/uid/generate`,
+        url: `/content-manager/uid/generate`,
         method: 'POST',
         body: {
           contentTypeUID: uid,
@@ -223,7 +223,7 @@ describe('Content Manager single types', () => {
       });
 
       const thridRes = await rq({
-        url: `/content-manager/explorer/uid/generate`,
+        url: `/content-manager/uid/generate`,
         method: 'POST',
         body: {
           contentTypeUID: uid,
@@ -242,7 +242,7 @@ describe('Content Manager single types', () => {
   describe('Check UID availability', () => {
     test('Throws if input is not provided', async () => {
       const res = await rq({
-        url: `/content-manager/explorer/uid/check-availability`,
+        url: `/content-manager/uid/check-availability`,
         method: 'POST',
         body: {},
       });
@@ -262,7 +262,7 @@ describe('Content Manager single types', () => {
 
     test('Throws on invalid uid value', async () => {
       const res = await rq({
-        url: `/content-manager/explorer/uid/check-availability`,
+        url: `/content-manager/uid/check-availability`,
         method: 'POST',
         body: {
           contentTypeUID: uid,
@@ -281,7 +281,7 @@ describe('Content Manager single types', () => {
 
     test('Throws when contentType is not found', async () => {
       const res = await rq({
-        url: `/content-manager/explorer/uid/check-availability`,
+        url: `/content-manager/uid/check-availability`,
         method: 'POST',
         body: {
           contentTypeUID: 'non-existent',
@@ -301,7 +301,7 @@ describe('Content Manager single types', () => {
 
     test('Throws when field is not a uid field', async () => {
       const res = await rq({
-        url: `/content-manager/explorer/uid/check-availability`,
+        url: `/content-manager/uid/check-availability`,
         method: 'POST',
         body: {
           contentTypeUID: uid,
@@ -323,7 +323,7 @@ describe('Content Manager single types', () => {
 
     test('Checks availability', async () => {
       const res = await rq({
-        url: `/content-manager/explorer/uid/check-availability`,
+        url: `/content-manager/uid/check-availability`,
         method: 'POST',
         body: {
           contentTypeUID: uid,
@@ -350,7 +350,7 @@ describe('Content Manager single types', () => {
       });
 
       const res = await rq({
-        url: `/content-manager/explorer/uid/check-availability`,
+        url: `/content-manager/uid/check-availability`,
         method: 'POST',
         body: {
           contentTypeUID: uid,
